refactor: remove unused
remove some functions and files.

Nddtfjiang <zhicheng.jiang@merico.dev>
