+++ plugins/gitlab/gitlab.go
@@ -181,126 +181,6 @@ func (plugin Gitlab) Execute(options map[string]interface{}, progress chan<- flo
 		}
 	}
 
-	progress <- 0.1
-	/*if err := tasks.CollectProject(ctx, projectIdInt, gitlabApiClient); err != nil {
-		return fmt.Errorf("could not collect projects: %v", err)
-	}
-	if tasksToRun["collectCommits"] {
-		progress <- 0.25
-		if err := tasks.CollectCommits(ctx, projectIdInt, gitlabApiClient); err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "collectCommits",
-				Message:     fmt.Errorf("could not collect commits: %v", err).Error(),
-			}
-		}
-	}
-	if tasksToRun["collectTags"] {
-		progress <- 0.3
-		if err := tasks.CollectTags(ctx, projectIdInt, gitlabApiClient); err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "collectTags",
-				Message:     fmt.Errorf("could not collect tags: %v", err).Error(),
-			}
-		}
-	}
-	if tasksToRun["collectMrs"] {
-		progress <- 0.35
-		mergeRequestErr := tasks.CollectMergeRequests(ctx, projectIdInt, gitlabApiClient)
-		if mergeRequestErr != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "collectMrs",
-				Message:     fmt.Errorf("could not collect merge requests: %v", mergeRequestErr).Error(),
-			}
-		}
-	}
-
-	if tasksToRun["collectMrNotes"] {
-		progress <- 0.4
-		err = tasks.CollectMergeRequestNotes(ctx, projectIdInt, rateLimitPerSecondInt, gitlabApiClient)
-		if err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "collectMrNotes",
-				Message:     fmt.Errorf("could not collect merge request notes: %v", err).Error(),
-			}
-		}
-	}
-	if tasksToRun["collectMrCommits"] {
-		progress <- 0.45
-		err = tasks.CollectMergeRequestCommits(ctx, projectIdInt, rateLimitPerSecondInt, gitlabApiClient)
-		if err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "collectMrCommits",
-				Message:     fmt.Errorf("could not collect merge request commits: %v", err).Error(),
-			}
-		}
-	}
-	if tasksToRun["enrichMrs"] {
-		progress <- 0.5
-		enrichErr := tasks.EnrichMergeRequests(ctx, projectIdInt)
-		if enrichErr != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "enrichMrs",
-				Message:     fmt.Errorf("could not enrich merge requests: %v", enrichErr).Error(),
-			}
-		}
-	}
-
-		if tasksToRun["collectPipelines"] {
-			progress <- 0.6
-			if err := tasks.CollectAllPipelines(projectIdInt, gitlabApiClient); err != nil {
-				return &errors.SubTaskError{
-					SubTaskName: "collectPipelines",
-					Message:     fmt.Errorf("could not collect pipelines: %v", err).Error(),
-				}
-			}
-			if err := tasks.CollectChildrenOnPipelines(projectIdInt, gitlabApiClient); err != nil {
-				return &errors.SubTaskError{
-					SubTaskName: "collectChildrenOnPipelines",
-					Message:     fmt.Errorf("could not collect children pipelines: %v", err).Error(),
-				}
-			}
-		}
-
-	if tasksToRun["convertProjects"] {
-		progress <- 0.7
-		err = tasks.ConvertProjects(ctx, projectIdInt)
-		if err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "convertProjects",
-				Message:     err.Error(),
-			}
-		}
-	}
-	if tasksToRun["convertMrs"] {
-		progress <- 0.75
-		err = tasks.ConvertMrs(ctx, projectIdInt)
-		if err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "convertMrs",
-				Message:     err.Error(),
-			}
-		}
-	}
-	if tasksToRun["convertCommits"] {
-		progress <- 0.8
-		err = tasks.ConvertCommits(projectIdInt)
-		if err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "convertCommits",
-				Message:     err.Error(),
-			}
-		}
-	}
-	if tasksToRun["convertNotes"] {
-		progress <- 0.9
-		err = tasks.ConvertNotes(ctx, projectIdInt)
-		if err != nil {
-			return &errors.SubTaskError{
-				SubTaskName: "convertNotes",
-				Message:     err.Error(),
-			}
-		}
-	}*/
 	progress <- 1
 	return nil
 }
