+++ plugins/gitlab/tasks/gitlab_api_note_converter.go
@@ -3,9 +3,12 @@ package tasks
 import (
 	"reflect"
 
+	"github.com/merico-dev/lake/models/domainlayer"
+	"github.com/merico-dev/lake/models/domainlayer/code"
 	"github.com/merico-dev/lake/models/domainlayer/didgen"
 	"github.com/merico-dev/lake/plugins/core"
 	"github.com/merico-dev/lake/plugins/gitlab/models"
+	gitlabModels "github.com/merico-dev/lake/plugins/gitlab/models"
 	"github.com/merico-dev/lake/plugins/helper"
 )
 
@@ -42,4 +45,20 @@ func ConvertApiNotes(taskCtx core.SubTaskContext) error {
 	}
 
 	return converter.Execute()
+}
+
+func convertToNoteModel(note *gitlabModels.GitlabMergeRequestNote, domainIdGeneratorNote *didgen.DomainIdGenerator) *code.Note {
+	domainNote := &code.Note{
+		DomainEntity: domainlayer.DomainEntity{
+			Id: domainIdGeneratorNote.Generate(note.GitlabId),
+		},
+		PrId:        uint64(note.MergeRequestId),
+		Type:        note.NoteableType,
+		Author:      note.AuthorUsername,
+		Body:        note.Body,
+		Resolvable:  note.Resolvable,
+		System:      note.System,
+		CreatedDate: note.GitlabCreatedAt,
+	}
+	return domainNote
 }
