+++ plugins/gitlab/tasks/gitlab_api_project_collector.go
@@ -7,6 +7,24 @@ import (
 
 const RAW_PROJECT_TABLE = "gitlab_api_project"
 
+type GitlabApiProject struct {
+	GitlabId          int    `json:"id"`
+	Name              string `josn:"name"`
+	Description       string `json:"description"`
+	DefaultBranch     string `json:"default_branch"`
+	PathWithNamespace string `json:"path_with_namespace"`
+	WebUrl            string `json:"web_url"`
+	CreatorId         int
+	Visibility        string            `json:"visibility"`
+	OpenIssuesCount   int               `json:"open_issues_count"`
+	StarCount         int               `json:"star_count"`
+	ForkedFromProject *GitlabApiProject `json:"forked_from_project"`
+	CreatedAt         core.Iso8601Time  `json:"created_at"`
+	LastActivityAt    *core.Iso8601Time `json:"last_activity_at"`
+}
+
+type GitlabApiProjectResponse GitlabApiProject
+
 func CollectApiProject(taskCtx core.SubTaskContext) error {
 	rawDataSubTaskArgs, data := CreateRawDataSubTaskArgs(taskCtx, RAW_PROJECT_TABLE)
